+++ src/includes/classes/Core/Utils/FileExt.php
@@ -5,16 +5,16 @@
  * @author @jaswrks
  * @copyright WebSharks™
  */
-declare (strict_types = 1);
+declare(strict_types=1);
 namespace WebSharks\Core\Classes\Core\Utils;
 
 use WebSharks\Core\Classes;
 use WebSharks\Core\Interfaces;
 use WebSharks\Core\Traits;
-#
+//
 use WebSharks\Core\Classes\Core\Error;
 use WebSharks\Core\Classes\Core\Base\Exception;
-#
+//
 use function assert as debug;
 use function get_defined_vars as vars;
 
@@ -39,18 +39,50 @@ class FileExt extends Classes\Core\Base\Core
         if (!$path) {
             return ''; // Not possible.
         }
-        if (mb_substr($path, -1) === '/') {
-            return ''; // Directory.
+        if (in_array(mb_substr($path, -1), ['/', '\\', DIRECTORY_SEPARATOR], true)) {
+            return ''; // Directory, not a file.
+        } elseif (!($basename = basename($path))) {
+            return ''; // Not possible.
+        } elseif (!($ext = mb_strrchr($basename, '.'))) {
+            return ''; // No extension.
         }
-        if (!($basename = basename($path))) {
-            return ''; // Nothing.
-        }
-        if (!($ext = mb_strrchr($basename, '.'))) {
-            return ''; // Nothing.
-        }
-        $ext = $this->c::mbLTrim($ext, '.');
-        $ext = mb_strtolower($ext);
+        $ext        = $this->c::mbLTrim($ext, '.');
+        return $ext = mb_strtolower($ext);
+    }
 
-        return $ext;
+    /**
+     * Set file extension.
+     *
+     * @since 17xxxx Extension utils.
+     *
+     * @param string $file Image file.
+     * @param string $ext  New extension.
+     *
+     * @return string File w/ extension.
+     */
+    public function set(string $file, string $ext): string
+    {
+        if (!$file || !$ext) {
+            return $file; // Not possible.
+        }
+        return $file = preg_replace('/\.[^.]+$/ui', '', $file).'.'.$ext;
+    }
+
+    /**
+     * Change file extension.
+     *
+     * @since 17xxxx Extension utils.
+     *
+     * @param string $file Image file.
+     * @param string $ext  New extension.
+     *
+     * @return string File w/ possible extension.
+     */
+    public function change(string $file, string $ext): string
+    {
+        if (!$file || !$ext) {
+            return $file; // Not possible.
+        }
+        return $file = preg_replace('/\.[^.]+$/ui', '.'.$ext, $file);
     }
 }